pyproject.toml
[project]
name = "amt-project"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "torch",
    "torchaudio",
    "lightning",
    "hydra-core",
    "omegaconf",
    "wandb",
    "optuna",
    "pretty_midi",
    "mir_eval",
    "pandas",
    "numpy",
    "matplotlib",
    "tqdm",
    "librosa",
    "optuna-integration[pytorch_lightning]",
    "torchcodec",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]